refactor(train): import styletts2 engines torch-free
The engine registry is imported eagerly and is meant to stay importable
without torch/pytorch_lightning (the other engines defer their heavy
imports behind TYPE_CHECKING + local imports). The styletts2 engine
modules imported torch/pl at module top and subclassed pl.LightningModule
there, so `from phoonnx_train.engines import get_engine` pulled torch —
breaking the torch-free contract.

Move the torch LightningModule/Dataset classes into the vendored
phoonnx_train.styletts2 package (aligner_module, pitch_module,
plbert_module, tts_module), mirroring how the glowtts/matcha/vits engines
keep their LightningModules in a sibling package. The engine modules keep
only the torch-free Config + Engine classes, defer torch behind
TYPE_CHECKING, and expose the moved classes lazily via module __getattr__.
The registry now imports under a torch/pl/torchaudio import blocker.
